requirements.txt
django>=5.0.0
python-dotenv>=1.0.0
python-decouple>=3.8
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
serpapi>=0.1.0
PyPDF2>=3.0.0
python-magic>=0.4.27
mistralai>=0.4.2
requests>=2.31.0
beautifulsoup4>=4.12.2
fpdf>=1.7.2
openai>=1.3.0
markdown>=3.5
pydantic>=2.4.2